feat: test des alert dans la page vehicule | documentation : composant modal.tsx

+++ src/features/vehicules/pages/Vehicles.tsx
@@ -1,5 +1,5 @@
 import React, {useEffect, useRef, useState} from "react";
-import {IonFab, IonFabButton, IonFabList, IonIcon, IonList} from "@ionic/react";
+import {IonFab, IonFabButton, IonFabList, IonIcon, IonList, useIonAlert} from "@ionic/react";
 
 import {PageLayout} from "../../../shared/PageLayout";
 import {Modal} from "../../../shared/Modal";
@@ -23,6 +23,8 @@ export const Vehicles = () => {
   const [vehiculeUnite, setVehiculeUnite] = useState<VehiculeType>();
   // filter coming from FabFilter component, sending state for it to update it
   const [filter, setFiler] = useState<string>("all");
+
+  const [presentAlert] = useIonAlert();
 
   useEffect(() => {
     fetchVehicules();
@@ -48,6 +50,11 @@ export const Vehicles = () => {
       vehiculeUnite?.immatriculation !== ""
     ) {
       dataService.postData(dataURL.vehicules, vehiculeUnite).then(() => fetchVehicules());
+      presentAlert({
+        header: `${vehiculeUnite.marque}, ${vehiculeUnite.modele}`,
+        message: "A bien été créé",
+        buttons: ["OK"],
+      });
     } else {
       console.log("incomplet");
     }
@@ -65,16 +72,45 @@ export const Vehicles = () => {
       },
     };
     dataService.postData(dataURL.locations, createLocationObj);
+    presentAlert({
+      header: `La location de a été crée`,
+      buttons: ["OK"],
+    });
   };
 
   // VEHICLE EDIT :
   const submitEditedElement = (obj: VehiculeType) => {
     dataService.putData(dataURL.vehicules, obj).then(() => fetchVehicules());
+    presentAlert({
+      header: `${obj.modele}, ${obj.marque}`,
+      message: "A bien été édité",
+      buttons: ["OK"],
+    });
   };
 
   // VEHICLE DELETION :
-  const deleteItem = (id: string) => {
-    dataService.deleteData(dataURL.vehicules, id).then(() => fetchVehicules());
+  const deleteElement = (id: string) => {
+    if (vehiculeList) {
+      let deletedElement = vehiculeList.filter((vehicule) => vehicule.id === id);
+
+      presentAlert({
+        header: `Voulez vous Supprimer : `,
+        message: `${deletedElement[0].modele}, ${deletedElement[0].marque}`,
+        buttons: [
+          {
+            text: "Annuler",
+            role: "cancel",
+          },
+          {
+            text: "Supprimer",
+            role: "confirm",
+            handler: () => {
+              dataService.deleteData(dataURL.vehicules, id).then(() => fetchVehicules());
+            },
+          },
+        ],
+      });
+    }
   };
 
   // DISPLAY
@@ -94,13 +130,13 @@ export const Vehicles = () => {
       <FabFilter setStateFilter={setFiler} />
 
       <Modal
-        modalTitle="Ajouter :"
-        triggerOpenModal="to-open-modal-vehicle"
-        formToDisplay="vehicule"
-        objectToManage={vehiculeUnite}
+        elementToManage={vehiculeUnite}
         handleInput={handleInput}
         submitNewElement={submitNewElement}
-        submitNewLocationElement={() => {}}></Modal>
+        submitNewLocationElement={() => {}}
+        modalTitle="Ajouter :"
+        triggerOpenModal="to-open-modal-vehicle"
+        formToDisplay="vehicule"></Modal>
 
       <IonList class="list-additional-style">
         {vehiculeList &&
@@ -108,13 +144,13 @@ export const Vehicles = () => {
             return (
               <CardLayout
                 key={vehicule.id}
-                elementType={vehicule}
+                elementToManage={vehicule}
                 submitEditedElement={submitEditedElement}
                 submitNewLocationElement={submitNewLocationElement}
-                deleteElement={deleteItem}
-                IsRental={true}
+                deleteElement={deleteElement}
                 triggerModalId={`to-edit-client${vehicule.id}`}
-                formType="vehicule">
+                formToDisplay="vehicule"
+                IsRental={true}>
                 <VehiculeCard element={vehicule} />
               </CardLayout>
             );
